// src/fizzy-effect/types.ts
export interface UpdateStepRequestContent {
	content?: string;
	completed?: boolean;
}